requirements.txt
tensorflow-gpu
keras==2.2.0
pillow==5.2.0
toposort==1.5
horovod==0.13.8
